+++ src/main/java/com/zipline/controller/agentProperty/AgentPropertyController.java
@@ -1,4 +1,4 @@
-package com.zipline.controller;
+package com.zipline.controller.agentProperty;
 
 import java.security.Principal;
 
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.zipline.dto.AgentPropertyRequestDTO;
-import com.zipline.dto.AgentPropertyResponseDTO;
+import com.zipline.dto.agentProperty.AgentPropertyRequestDTO;
+import com.zipline.dto.agentProperty.AgentPropertyResponseDTO;
 import com.zipline.global.common.response.ApiResponse;
-import com.zipline.service.AgentPropertyService;
+import com.zipline.service.agentProperty.AgentPropertyService;
 
 import lombok.RequiredArgsConstructor;
