+++ src/main/java/com/zipline/service/agentProperty/AgentPropertyService.java
@@ -1,4 +1,4 @@
-package com.zipline.service;
+package com.zipline.service.agentProperty;
 
 import java.time.LocalDateTime;
 
@@ -6,18 +6,18 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.zipline.dto.AgentPropertyRequestDTO;
-import com.zipline.dto.AgentPropertyResponseDTO;
-import com.zipline.entity.AgentProperty;
+import com.zipline.dto.agentProperty.AgentPropertyRequestDTO;
+import com.zipline.dto.agentProperty.AgentPropertyResponseDTO;
 import com.zipline.entity.Customer;
 import com.zipline.entity.User;
+import com.zipline.entity.agentProperty.AgentProperty;
 import com.zipline.global.exception.custom.PermissionDeniedException;
 import com.zipline.global.exception.custom.UserNotFoundException;
 import com.zipline.global.exception.custom.agentProperty.PropertyNotFoundException;
 import com.zipline.global.exception.custom.customer.CustomerNotFoundException;
-import com.zipline.repository.AgentPropertyRepository;
 import com.zipline.repository.CustomerRepository;
 import com.zipline.repository.UserRepository;
+import com.zipline.repository.agentProperty.AgentPropertyRepository;
 
 import lombok.RequiredArgsConstructor;
